id changed to long

+++ src/main/java/com/shymoniak/hospital/entity/Diagnosis.java
@@ -18,7 +18,7 @@ public class Diagnosis {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "diagnosis_id", nullable = false)
-    private int diagnosisId;
+    private long diagnosisId;
 
     @Column(name = "diagnosis_name", nullable = false, length = 100)
     private String name;
